add generic output plugin which simplifies some of the output plugin construction
